Fixed IO + added armSimConfig
